@@ -1,6 +1,6 @@
 {
   "name": "react-color",
-  "version": "1.0.3",
+  "version": "2.1.0",
   "description": "A Collection of Color Pickers from Sketch, Photoshop, Chrome & more",
   "main": "lib/index.js",
   "repository": {
@@ -11,6 +11,14 @@
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/casesandberg/react-color/issues"
+  },
+  "scripts": {
+    "unit-test": "./node_modules/mocha/bin/mocha \"test/**/*.test.js\" --compilers js:babel-register",
+    "test": "npm run unit-test",
+    "clean": "rm -rf lib && mkdir lib",
+    "lib": "npm run clean && cp -a src/. lib && node map-styles && ./node_modules/.bin/babel lib -d lib",
+    "docs": "node docs-server",
+    "docs-dist": "node docs-dist"
   },
   "homepage": "http://casesandberg.github.io/react-color/",
   "keywords": [
@@ -26,33 +34,47 @@
     "popup"
   ],
   "dependencies": {
-    "install": "^0.1.8",
-    "lodash": "^3.10.1",
+    "lodash": "^4.0.1",
+    "lodash.debounce": "^4.0.0",
+    "lodash.isplainobject": "^4.0.0",
+    "lodash.throttle": "^4.0.0",
     "material-colors": "^1.0.0",
     "merge": "^1.2.0",
-    "npm": "^2.13.4",
-    "react": "^0.13.3",
-    "reactcss": "^0.3.2",
+    "react-addons-shallow-compare": "^15.0.0",
+    "reactcss": "^0.4.3",
     "tinycolor2": "^1.1.2"
   },
+  "peerDependencies": {
+    "react": "^0.14.0 || ^15.0.0",
+    "react-dom": "^0.14.0 || ^15.0.0"
+  },
   "devDependencies": {
-    "babel": "^5.8.21",
-    "babel-core": "^5.8.22",
-    "babel-loader": "^5.3.2",
-    "css-loader": "^0.15.6",
-    "gulp": "^3.9.0",
-    "gulp-babel": "^5.2.0",
-    "gulp-modify": "^0.1.1",
-    "gulp-util": "^3.0.6",
-    "highlight.js": "^8.7.0",
+    "babel-cli": "^6.8.0",
+    "babel-core": "^6.4.0",
+    "babel-loader": "^6.2.1",
+    "babel-plugin-lodash": "^3.1.0",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-0": "^6.3.13",
+    "babel-register": "^6.5.2",
+    "chai": "^3.3.0",
+    "chai-spies": "^0.7.1",
+    "css-loader": "^0.23.1",
+    "event-stream": "^3.3.1",
+    "highlight.js": "^9.3.0",
     "html-loader": "^0.3.0",
     "jsx-loader": "^0.13.2",
-    "normalize.css": "^3.0.3",
-    "react-context": "0.0.1",
+    "lodash": "^4.6.0",
+    "normalize.css": "^4.1.1",
+    "react-addons-test-utils": "^15.0.0",
+    "react-context": "0.0.3",
     "react-hot-loader": "^1.2.8",
-    "react-map-styles": "^0.2.0",
+    "react-map-styles": "^0.3.0",
     "remarkable": "^1.6.0",
-    "style-loader": "^0.12.3",
+    "require-dir": "^0.3.0",
+    "rimraf": "^2.5.0",
+    "style-loader": "^0.13.0",
+    "testdom": "^2.0.0",
     "webpack": "^1.11.0",
     "webpack-dev-server": "^1.10.1"
   }
