@@ -1,17 +1,18 @@
-'use strict';
+'use strict'
 
-var React = require('react');
+import React from 'react'
+import ReactDOM from 'react-dom'
 
-var Home = require('./components/home/Home');
+import Home from './components/home/Home'
 
-// var html = React.renderToString(React.createElement(Home));
+// var html = ReactDOMServer.renderToString(React.createElement(Home));
 // console.log(html);
 
 if (typeof document !== 'undefined') {
-  React.render(
+  ReactDOM.render(
     React.createElement(Home),
     document.getElementById('root')
-  );
+  )
 }
 
-module.exports = Home;
+module.exports = Home
