@@ -1,2 +1,12 @@
+'use strict'
 
-module.exports = require('./components/Color');
+export { default as ChromePicker } from './components/chrome/Chrome'
+export { default as CompactPicker } from './components/compact/Compact'
+export { default as MaterialPicker } from './components/material/Material'
+export { default as PhotoshopPicker } from './components/photoshop/Photoshop'
+export { default as SketchPicker } from './components/sketched/Sketch'
+export { default as SliderPicker } from './components/slider/Slider'
+export { default as SwatchesPicker } from './components/swatches/Swatches'
+export { default as CustomPicker } from './components/common/ColorWrap'
+
+export default from './components/chrome/Chrome'
