@@ -4,18 +4,12 @@ module.exports = {
   '02-getting-started': require('./02-getting-started.md'),
   '02.01-install': require('./02.01-install.md'),
   '02.02-include': require('./02.02-include.md'),
-  '02.03-display': require('./02.03-display.md'),
   '03-api': require('./03-api.md'),
-  '03.01-type': require('./03.01-type.md'),
+  '03.01-color': require('./03.01-color.md'),
   '03.02-onChange': require('./03.02-onChange.md'),
   '03.03-onChangeComplete': require('./03.03-onChangeComplete.md'),
-  '03.04-color': require('./03.04-color.md'),
-  '03.05-display': require('./03.05-display.md'),
-  '03.06-onClose': require('./03.06-onClose.md'),
-  '03.07-position': require('./03.07-position.md'),
-  '03.08-positionCSS': require('./03.08-positionCSS.md'),
   '04-create': require('./04-create.md'),
   '04.01-parent': require('./04.01-parent.md'),
   '04.02-helpers': require('./04.02-helpers.md'),
   '05-examples': require('./05-examples.md'),
-};
+}
