@@ -1,38 +1,64 @@
 'use strict';
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.Chrome = undefined;
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; desc = parent = getter = undefined; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+var _react = require('react');
 
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
+var _react2 = _interopRequireDefault(_react);
 
-function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+var _reactcss = require('reactcss');
 
-var React = require('react');
-var ReactCSS = require('reactcss');
+var _reactcss2 = _interopRequireDefault(_reactcss);
 
-var _require = require('../common');
+var _common = require('../common');
 
-var Saturation = _require.Saturation;
-var Hue = _require.Hue;
-var Alpha = _require.Alpha;
-var Checkboard = _require.Checkboard;
+var _ChromeFields = require('./ChromeFields');
 
-var ChromeFields = require('./ChromeFields');
-var ChromePointer = require('./ChromePointer');
-var ChromePointerCircle = require('./ChromePointerCircle');
+var _ChromeFields2 = _interopRequireDefault(_ChromeFields);
 
-var Chrome = (function (_ReactCSS$Component) {
+var _ChromePointer = require('./ChromePointer');
+
+var _ChromePointer2 = _interopRequireDefault(_ChromePointer);
+
+var _ChromePointerCircle = require('./ChromePointerCircle');
+
+var _ChromePointerCircle2 = _interopRequireDefault(_ChromePointerCircle);
+
+var _reactAddonsShallowCompare = require('react-addons-shallow-compare');
+
+var _reactAddonsShallowCompare2 = _interopRequireDefault(_reactAddonsShallowCompare);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var Chrome = exports.Chrome = function (_ReactCSS$Component) {
   _inherits(Chrome, _ReactCSS$Component);
 
   function Chrome() {
+    var _Object$getPrototypeO;
+
+    var _temp, _this, _ret;
+
+    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
+      args[_key] = arguments[_key];
+    }
+
     _classCallCheck(this, Chrome);
 
-    _get(Object.getPrototypeOf(Chrome.prototype), 'constructor', this).call(this);
-
-    this.handleChange = this.handleChange.bind(this);
+    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(Chrome)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.shouldComponentUpdate = _reactAddonsShallowCompare2.default.bind(_this, _this, arguments[0], arguments[1]), _this.handleChange = function (data) {
+      _this.props.onChange(data);
+    }, _temp), _possibleConstructorReturn(_this, _ret);
   }
 
   _createClass(Chrome, [{
@@ -44,6 +70,7 @@ var Chrome = (function (_ReactCSS$Component) {
             background: '#fff',
             borderRadius: '2px',
             boxShadow: '0 0 2px rgba(0,0,0,.3), 0 4px 8px rgba(0,0,0,.3)',
+            boxSizing: 'initial',
             width: '225px',
             fontFamily: 'Menlo'
           },
@@ -76,7 +103,6 @@ var Chrome = (function (_ReactCSS$Component) {
           },
           active: {
             Absolute: '0 0 0 0',
-            zIndex: 2,
             borderRadius: '8px',
             boxShadow: 'inset 0 0 0 1px rgba(0,0,0,.1)',
             background: 'rgba(' + this.props.rgb.r + ', ' + this.props.rgb.g + ', ' + this.props.rgb.b + ', ' + this.props.rgb.a + ')'
@@ -103,59 +129,54 @@ var Chrome = (function (_ReactCSS$Component) {
       };
     }
   }, {
-    key: 'handleChange',
-    value: function handleChange(data) {
-      this.props.onChange(data);
-    }
-  }, {
     key: 'render',
     value: function render() {
-      return React.createElement(
+      return _react2.default.createElement(
         'div',
         { style: this.styles().picker },
-        React.createElement(
+        _react2.default.createElement(
           'div',
           { style: this.styles().saturation },
-          React.createElement(Saturation, _extends({}, this.styles().Saturation, this.props, { pointer: ChromePointerCircle, onChange: this.handleChange }))
+          _react2.default.createElement(_common.Saturation, _extends({}, this.styles().Saturation, this.props, { pointer: _ChromePointerCircle2.default, onChange: this.handleChange }))
         ),
-        React.createElement(
+        _react2.default.createElement(
           'div',
           { style: this.styles().body },
-          React.createElement(
+          _react2.default.createElement(
             'div',
             { style: this.styles().controls, className: 'flexbox-fix' },
-            React.createElement(
+            _react2.default.createElement(
               'div',
               { style: this.styles().color },
-              React.createElement(
+              _react2.default.createElement(
                 'div',
                 { style: this.styles().swatch },
-                React.createElement('div', { style: this.styles().active }),
-                React.createElement(Checkboard, null)
+                _react2.default.createElement('div', { style: this.styles().active }),
+                _react2.default.createElement(_common.Checkboard, null)
               )
             ),
-            React.createElement(
+            _react2.default.createElement(
               'div',
               { style: this.styles().toggles },
-              React.createElement(
+              _react2.default.createElement(
                 'div',
                 { style: this.styles().hue },
-                React.createElement(Hue, _extends({}, this.styles().Hue, this.props, { pointer: ChromePointer, onChange: this.handleChange }))
+                _react2.default.createElement(_common.Hue, _extends({}, this.styles().Hue, this.props, { pointer: _ChromePointer2.default, onChange: this.handleChange }))
               ),
-              React.createElement(
+              _react2.default.createElement(
                 'div',
                 { style: this.styles().alpha },
-                React.createElement(Alpha, _extends({}, this.styles().Alpha, this.props, { pointer: ChromePointer, onChange: this.handleChange }))
+                _react2.default.createElement(_common.Alpha, _extends({}, this.styles().Alpha, this.props, { pointer: _ChromePointer2.default, onChange: this.handleChange }))
               )
             )
           ),
-          React.createElement(ChromeFields, _extends({}, this.props, { onChange: this.handleChange }))
+          _react2.default.createElement(_ChromeFields2.default, _extends({}, this.props, { onChange: this.handleChange }))
         )
       );
     }
   }]);
 
   return Chrome;
-})(ReactCSS.Component);
+}(_reactcss2.default.Component);
 
-module.exports = Chrome;
+exports.default = (0, _common.ColorWrap)(Chrome);
