@@ -1,29 +1,84 @@
 'use strict';
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.Alpha = undefined;
 
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; desc = parent = getter = undefined; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+var _react = require('react');
 
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
+var _react2 = _interopRequireDefault(_react);
 
-function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+var _reactcss = require('reactcss');
 
-var React = require('react');
-var ReactCSS = require('reactcss');
+var _reactcss2 = _interopRequireDefault(_reactcss);
 
-var Checkboard = require('./Checkboard');
+var _reactAddonsShallowCompare = require('react-addons-shallow-compare');
 
-var Alpha = (function (_ReactCSS$Component) {
+var _reactAddonsShallowCompare2 = _interopRequireDefault(_reactAddonsShallowCompare);
+
+var _Checkboard = require('./Checkboard');
+
+var _Checkboard2 = _interopRequireDefault(_Checkboard);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var Alpha = exports.Alpha = function (_ReactCSS$Component) {
   _inherits(Alpha, _ReactCSS$Component);
 
   function Alpha() {
+    var _Object$getPrototypeO;
+
+    var _temp, _this, _ret;
+
+    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
+      args[_key] = arguments[_key];
+    }
+
     _classCallCheck(this, Alpha);
 
-    _get(Object.getPrototypeOf(Alpha.prototype), 'constructor', this).call(this);
+    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(Alpha)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.shouldComponentUpdate = _reactAddonsShallowCompare2.default.bind(_this, _this, arguments[0], arguments[1]), _this.handleChange = function (e, skip) {
+      !skip && e.preventDefault();
+      var container = _this.refs.container;
+      var containerWidth = container.clientWidth;
+      var x = typeof e.pageX === 'number' ? e.pageX : e.touches[0].pageX;
+      var left = x - (container.getBoundingClientRect().left + window.pageXOffset);
 
-    this.handleChange = this.handleChange.bind(this);
-    this.handleMouseDown = this.handleMouseDown.bind(this);
-    this.handleMouseUp = this.handleMouseUp.bind(this);
+      var a;
+      if (left < 0) {
+        a = 0;
+      } else if (left > containerWidth) {
+        a = 1;
+      } else {
+        a = Math.round(left * 100 / containerWidth) / 100;
+      }
+
+      if (_this.props.a !== a) {
+        _this.props.onChange({
+          h: _this.props.hsl.h,
+          s: _this.props.hsl.s,
+          l: _this.props.hsl.l,
+          a: a,
+          source: 'rgb'
+        });
+      }
+    }, _this.handleMouseDown = function (e) {
+      _this.handleChange(e, true);
+      window.addEventListener('mousemove', _this.handleChange);
+      window.addEventListener('mouseup', _this.handleMouseUp);
+    }, _this.handleMouseUp = function () {
+      _this.unbindEventListeners();
+    }, _this.unbindEventListeners = function () {
+      window.removeEventListener('mousemove', _this.handleChange);
+      window.removeEventListener('mouseup', _this.handleMouseUp);
+    }, _temp), _possibleConstructorReturn(_this, _ret);
   }
 
   _createClass(Alpha, [{
@@ -47,12 +102,10 @@ var Alpha = (function (_ReactCSS$Component) {
           },
           container: {
             position: 'relative',
-            zIndex: '2',
             height: '100%',
             margin: '0 3px'
           },
           pointer: {
-            zIndex: '2',
             position: 'absolute',
             left: this.props.rgb.a * 100 + '%'
           },
@@ -69,61 +122,34 @@ var Alpha = (function (_ReactCSS$Component) {
       };
     }
   }, {
-    key: 'handleChange',
-    value: function handleChange(e) {
-      var container = React.findDOMNode(this.refs.container);
-      var containerWidth = container.clientWidth;
-      var left = e.pageX - (container.getBoundingClientRect().left + window.pageXOffset);
-
-      var a;
-      if (left < 0) {
-        a = 0;
-      } else if (left > containerWidth) {
-        a = 1;
-      } else {
-        a = Math.round(left * 100 / containerWidth) / 100;
-      }
-
-      if (this.props.a !== a) {
-        this.props.onChange({ h: this.props.hsl.h, s: this.props.hsl.s, l: this.props.hsl.l, a: a });
-      }
-    }
-  }, {
-    key: 'handleMouseDown',
-    value: function handleMouseDown(e) {
-      e.preventDefault();
-      this.handleChange(e);
-      window.addEventListener('mousemove', this.handleChange);
-      window.addEventListener('mouseup', this.handleMouseUp);
-    }
-  }, {
-    key: 'handleMouseUp',
-    value: function handleMouseUp() {
-      window.removeEventListener('mousemove', this.handleChange);
-      window.removeEventListener('mouseup', this.handleMouseUp);
+    key: 'componentWillUnmount',
+    value: function componentWillUnmount() {
+      this.unbindEventListeners();
     }
   }, {
     key: 'render',
     value: function render() {
-      var pointer = React.createElement('div', { style: this.styles().slider });
+      var pointer = _react2.default.createElement('div', { style: this.styles().slider });
 
       if (this.props.pointer) {
-        pointer = React.createElement(this.props.pointer, this.props);
+        pointer = _react2.default.createElement(this.props.pointer, this.props);
       }
 
-      return React.createElement(
+      return _react2.default.createElement(
         'div',
         { style: this.styles().alpha },
-        React.createElement(
+        _react2.default.createElement(
           'div',
           { style: this.styles().checkboard },
-          React.createElement(Checkboard, null)
+          _react2.default.createElement(_Checkboard2.default, null)
         ),
-        React.createElement('div', { style: this.styles().gradient }),
-        React.createElement(
+        _react2.default.createElement('div', { style: this.styles().gradient }),
+        _react2.default.createElement(
           'div',
-          { style: this.styles().container, ref: 'container', onMouseDown: this.handleMouseDown },
-          React.createElement(
+          { style: this.styles().container, ref: 'container', onMouseDown: this.handleMouseDown,
+            onTouchMove: this.handleChange,
+            onTouchStart: this.handleChange },
+          _react2.default.createElement(
             'div',
             { style: this.styles().pointer, ref: 'pointer' },
             pointer
@@ -134,6 +160,6 @@ var Alpha = (function (_ReactCSS$Component) {
   }]);
 
   return Alpha;
-})(ReactCSS.Component);
+}(_reactcss2.default.Component);
 
-module.exports = Alpha;
+exports.default = Alpha;
