@@ -1,103 +1,104 @@
 'use strict';
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.Checkboard = undefined;
 
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; desc = parent = getter = undefined; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+var _react = require('react');
 
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
+var _react2 = _interopRequireDefault(_react);
 
-function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+var _reactcss = require('reactcss');
 
-var React = require('react');
-var ReactCSS = require('reactcss');
+var _reactcss2 = _interopRequireDefault(_reactcss);
 
-var Checkboard = (function (_ReactCSS$Component) {
+var _reactAddonsShallowCompare = require('react-addons-shallow-compare');
+
+var _reactAddonsShallowCompare2 = _interopRequireDefault(_reactAddonsShallowCompare);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var _checkboardCache = {};
+
+function renderCheckboard(c1, c2, size) {
+  if (typeof document == 'undefined') return null; // Dont Render On Server
+  var canvas = document.createElement('canvas');
+  canvas.width = canvas.height = size * 2;
+  var ctx = canvas.getContext('2d');
+  if (!ctx) return null; // If no context can be found, return early.
+  ctx.fillStyle = c1;
+  ctx.fillRect(0, 0, canvas.width, canvas.height);
+  ctx.fillStyle = c2;
+  ctx.fillRect(0, 0, size, size);
+  ctx.translate(size, size);
+  ctx.fillRect(0, 0, size, size);
+  return canvas.toDataURL();
+}
+
+function getCheckboard(c1, c2, size) {
+  var key = c1 + ',' + c2 + ',' + size;
+
+  if (_checkboardCache[key]) {
+    return _checkboardCache[key];
+  } else {
+    var checkboard = renderCheckboard(c1, c2, size);
+    _checkboardCache[key] = checkboard;
+    return checkboard;
+  }
+}
+
+var Checkboard = exports.Checkboard = function (_ReactCSS$Component) {
   _inherits(Checkboard, _ReactCSS$Component);
 
   function Checkboard() {
+    var _Object$getPrototypeO;
+
+    var _temp, _this, _ret;
+
+    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
+      args[_key] = arguments[_key];
+    }
+
     _classCallCheck(this, Checkboard);
 
-    _get(Object.getPrototypeOf(Checkboard.prototype), 'constructor', this).call(this);
-
-    this.state = {
-      children: []
-    };
+    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(Checkboard)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.shouldComponentUpdate = _reactAddonsShallowCompare2.default.bind(_this, _this, arguments[0], arguments[1]), _temp), _possibleConstructorReturn(_this, _ret);
   }
 
   _createClass(Checkboard, [{
     key: 'classes',
     value: function classes() {
+      var background = getCheckboard(this.props.white, this.props.grey, this.props.size);
       return {
         'default': {
           grid: {
-            Absolute: '0 0 0 0'
-          },
-          inside: {
-            position: 'absolute',
-            left: '50%',
-            top: '50%',
-            transform: 'translate(-50%, -50%)'
-          },
-          white: {
-            width: this.props.size + 'px',
-            height: this.props.size + 'px',
-            background: '#fff',
-            float: 'left'
-          },
-          grey: {
-            width: this.props.size + 'px',
-            height: this.props.size + 'px',
-            background: '#e6e6e6',
-            float: 'left'
+            Absolute: '0 0 0 0',
+            background: 'url(' + background + ') center left'
           }
         }
       };
     }
   }, {
-    key: 'componentDidMount',
-    value: function componentDidMount() {
-      var grid = React.findDOMNode(this.refs.grid);
-      var inside = React.findDOMNode(this.refs.inside);
-      var rows = Math.ceil(grid.clientHeight / this.props.size);
-      var columns = Math.ceil(grid.clientWidth / this.props.size);
-
-      if (columns % 2 == 0) {
-        columns++;
-      }
-
-      inside.style.width = columns * this.props.size + 'px';
-
-      var children = [];
-      for (var i = 0; i < rows * columns; i++) {
-        if (i % 2 == 0) {
-          children.push(React.createElement('div', { key: i, style: this.styles().white }));
-        } else {
-          children.push(React.createElement('div', { key: i, style: this.styles().grey }));
-        }
-      }
-
-      this.setState({ children: children });
-    }
-  }, {
     key: 'render',
     value: function render() {
-      return React.createElement(
-        'div',
-        { style: this.styles().grid, ref: 'grid' },
-        React.createElement(
-          'div',
-          { style: this.styles().inside, ref: 'inside' },
-          this.state.children
-        )
-      );
+      return _react2.default.createElement('div', { style: this.styles().grid, ref: 'grid' });
     }
   }]);
 
   return Checkboard;
-})(ReactCSS.Component);
+}(_reactcss2.default.Component);
 
 Checkboard.defaultProps = {
-  size: 8
+  size: 8,
+  white: '#fff',
+  grey: '#e6e6e6'
 };
 
-module.exports = Checkboard;
+exports.default = Checkboard;
