@@ -1,32 +1,126 @@
 'use strict';
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.EditableInput = undefined;
 
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; desc = parent = getter = undefined; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+var _react = require('react');
 
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
+var _react2 = _interopRequireDefault(_react);
 
-function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+var _reactcss = require('reactcss');
 
-var React = require('react');
-var ReactCSS = require('reactcss');
+var _reactcss2 = _interopRequireDefault(_reactcss);
 
-var EditableInput = (function (_ReactCSS$Component) {
+var _reactAddonsShallowCompare = require('react-addons-shallow-compare');
+
+var _reactAddonsShallowCompare2 = _interopRequireDefault(_reactAddonsShallowCompare);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var EditableInput = exports.EditableInput = function (_ReactCSS$Component) {
   _inherits(EditableInput, _ReactCSS$Component);
 
   function EditableInput(props) {
     _classCallCheck(this, EditableInput);
 
-    _get(Object.getPrototypeOf(EditableInput.prototype), 'constructor', this).call(this);
+    var _this = _possibleConstructorReturn(this, Object.getPrototypeOf(EditableInput).call(this));
 
-    this.state = {
-      value: String(props.value).toUpperCase()
-    }, this.handleChange = this.handleChange.bind(this);
-    this.handleDrag = this.handleDrag.bind(this);
-    this.handleBlur = this.handleBlur.bind(this);
-    this.handleKeyDown = this.handleKeyDown.bind(this);
-    this.handleMouseDown = this.handleMouseDown.bind(this);
-    this.handleMouseUp = this.handleMouseUp.bind(this);
+    _this.shouldComponentUpdate = _reactAddonsShallowCompare2.default.bind(_this, _this, arguments[0], arguments[1]);
+
+    _this.handleBlur = function () {
+      if (_this.state.blurValue) {
+        _this.setState({ value: _this.state.blurValue, blurValue: null });
+      }
+    };
+
+    _this.handleChange = function (e) {
+      if (_this.props.label !== null) {
+        var obj = {};
+        obj[_this.props.label] = e.target.value;
+        _this.props.onChange(obj);
+      } else {
+        _this.props.onChange(e.target.value);
+      }
+
+      _this.setState({ value: e.target.value });
+    };
+
+    _this.handleKeyDown = function (e) {
+      var number = Number(e.target.value);
+      if (number) {
+        var amount = _this.props.arrowOffset || 1;
+
+        // Up
+        if (e.keyCode === 38) {
+          if (_this.props.label !== null) {
+            var obj = {};
+            obj[_this.props.label] = number + amount;
+            _this.props.onChange(obj);
+          } else {
+            _this.props.onChange(number + amount);
+          }
+
+          _this.setState({ value: number + amount });
+        }
+
+        // Down
+        if (e.keyCode === 40) {
+          if (_this.props.label !== null) {
+            var obj = {};
+            obj[_this.props.label] = number - amount;
+            _this.props.onChange(obj);
+          } else {
+            _this.props.onChange(number - amount);
+          }
+
+          _this.setState({ value: number - amount });
+        }
+      }
+    };
+
+    _this.handleDrag = function (e) {
+      if (_this.props.dragLabel) {
+        var newValue = Math.round(_this.props.value + e.movementX);
+        if (newValue >= 0 && newValue <= _this.props.dragMax) {
+          var obj = {};
+          obj[_this.props.label] = newValue;
+          _this.props.onChange(obj);
+        }
+      }
+    };
+
+    _this.handleMouseDown = function (e) {
+      if (_this.props.dragLabel) {
+        e.preventDefault();
+        _this.handleDrag(e);
+        window.addEventListener('mousemove', _this.handleDrag);
+        window.addEventListener('mouseup', _this.handleMouseUp);
+      }
+    };
+
+    _this.handleMouseUp = function () {
+      _this.unbindEventListeners();
+    };
+
+    _this.unbindEventListeners = function () {
+      window.removeEventListener('mousemove', _this.handleChange);
+      window.removeEventListener('mouseup', _this.handleMouseUp);
+    };
+
+    _this.state = {
+      value: String(props.value).toUpperCase(),
+      blurValue: String(props.value).toUpperCase()
+    };
+    return _this;
   }
 
   _createClass(EditableInput, [{
@@ -55,7 +149,7 @@ var EditableInput = (function (_ReactCSS$Component) {
   }, {
     key: 'componentWillReceiveProps',
     value: function componentWillReceiveProps(nextProps) {
-      var input = React.findDOMNode(this.refs.input);
+      var input = this.refs.input;
       if (nextProps.value !== this.state.value) {
         if (input === document.activeElement) {
           this.setState({ blurValue: String(nextProps.value).toUpperCase() });
@@ -65,109 +159,32 @@ var EditableInput = (function (_ReactCSS$Component) {
       }
     }
   }, {
-    key: 'handleBlur',
-    value: function handleBlur() {
-      if (this.state.blurValue) {
-        this.setState({ value: this.state.blurValue, blurValue: null });
-      }
-    }
-  }, {
-    key: 'handleChange',
-    value: function handleChange(e) {
-      if (this.props.label !== null) {
-        var obj = {};
-        obj[this.props.label] = e.target.value;
-        this.props.onChange(obj);
-      } else {
-        this.props.onChange(e.target.value);
-      }
-
-      this.setState({ value: e.target.value });
-    }
-  }, {
-    key: 'handleKeyDown',
-    value: function handleKeyDown(e) {
-      var number = Number(e.target.value);
-      if (number) {
-        var amount = this.props.arrowOffset || 1;
-
-        // Up
-        if (e.keyCode === 38) {
-          if (this.props.label !== null) {
-            var obj = {};
-            obj[this.props.label] = number + amount;
-            this.props.onChange(obj);
-          } else {
-            this.props.onChange(number + amount);
-          }
-
-          this.setState({ value: number + amount });
-        }
-
-        // Down
-        if (e.keyCode === 40) {
-          if (this.props.label !== null) {
-            var obj = {};
-            obj[this.props.label] = number - amount;
-            this.props.onChange(obj);
-          } else {
-            this.props.onChange(number - amount);
-          }
-
-          this.setState({ value: number - amount });
-        }
-      }
-    }
-  }, {
-    key: 'handleDrag',
-    value: function handleDrag(e) {
-      if (this.props.dragLabel) {
-        var newValue = Math.round(this.props.value + e.movementX);
-        if (newValue >= 0 && newValue <= this.props.dragMax) {
-          var obj = {};
-          obj[this.props.label] = newValue;
-          this.props.onChange(obj);
-        }
-      }
-    }
-  }, {
-    key: 'handleMouseDown',
-    value: function handleMouseDown(e) {
-      if (this.props.dragLabel) {
-        e.preventDefault();
-        this.handleDrag(e);
-        window.addEventListener('mousemove', this.handleDrag);
-        window.addEventListener('mouseup', this.handleMouseUp);
-      }
-    }
-  }, {
-    key: 'handleMouseUp',
-    value: function handleMouseUp() {
-      window.removeEventListener('mousemove', this.handleDrag);
-      window.removeEventListener('mouseup', this.handleMouseUp);
+    key: 'componentWillUnmount',
+    value: function componentWillUnmount() {
+      this.unbindEventListeners();
     }
   }, {
     key: 'render',
     value: function render() {
       var label;
       if (this.props.label) {
-        label = React.createElement(
+        label = _react2.default.createElement(
           'span',
           { style: this.styles().label, ref: 'label', onMouseDown: this.handleMouseDown },
           this.props.label
         );
       }
 
-      return React.createElement(
+      return _react2.default.createElement(
         'div',
         { style: this.styles().wrap, ref: 'container' },
-        React.createElement('input', { style: this.styles().input, ref: 'input', value: this.state.value, onKeyDown: this.handleKeyDown, onChange: this.handleChange, onBlur: this.handleBlur }),
+        _react2.default.createElement('input', { style: this.styles().input, ref: 'input', value: this.state.value, onKeyDown: this.handleKeyDown, onChange: this.handleChange, onBlur: this.handleBlur }),
         label
       );
     }
   }]);
 
   return EditableInput;
-})(ReactCSS.Component);
+}(_reactcss2.default.Component);
 
-module.exports = EditableInput;
+exports.default = EditableInput;
