@@ -1,32 +1,91 @@
 'use strict';
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.Saturation = undefined;
 
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; desc = parent = getter = undefined; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+var _react = require('react');
 
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
+var _react2 = _interopRequireDefault(_react);
 
-function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+var _reactcss = require('reactcss');
 
-var React = require('react');
-var ReactCSS = require('reactcss');
-var _ = require('lodash');
+var _reactcss2 = _interopRequireDefault(_reactcss);
 
-var Saturation = (function (_ReactCSS$Component) {
+var _lodash = require('lodash.throttle');
+
+var _lodash2 = _interopRequireDefault(_lodash);
+
+var _reactAddonsShallowCompare = require('react-addons-shallow-compare');
+
+var _reactAddonsShallowCompare2 = _interopRequireDefault(_reactAddonsShallowCompare);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var Saturation = exports.Saturation = function (_ReactCSS$Component) {
   _inherits(Saturation, _ReactCSS$Component);
 
   function Saturation(props) {
     _classCallCheck(this, Saturation);
 
-    _get(Object.getPrototypeOf(Saturation.prototype), 'constructor', this).call(this);
+    var _this = _possibleConstructorReturn(this, Object.getPrototypeOf(Saturation).call(this));
 
-    this.throttle = _.throttle(function (fn, data) {
+    _this.shouldComponentUpdate = _reactAddonsShallowCompare2.default.bind(_this, _this, arguments[0], arguments[1]);
+
+    _this.handleChange = function (e, skip) {
+      !skip && e.preventDefault();
+      var container = _this.refs.container;
+      var containerWidth = container.clientWidth;
+      var containerHeight = container.clientHeight;
+      var x = typeof e.pageX === 'number' ? e.pageX : e.touches[0].pageX;
+      var y = typeof e.pageY === 'number' ? e.pageY : e.touches[0].pageY;
+      var left = x - (container.getBoundingClientRect().left + window.pageXOffset);
+      var top = y - (container.getBoundingClientRect().top + window.pageYOffset);
+
+      if (left < 0) {
+        left = 0;
+      } else if (left > containerWidth) {
+        left = containerWidth;
+      } else if (top < 0) {
+        top = 0;
+      } else if (top > containerHeight) {
+        top = containerHeight;
+      }
+
+      var saturation = left * 100 / containerWidth;
+      var bright = -(top * 100 / containerHeight) + 100;
+
+      _this.throttle(_this.props.onChange, {
+        h: _this.props.hsl.h,
+        s: saturation,
+        v: bright,
+        a: _this.props.hsl.a,
+        source: 'rgb'
+      });
+    };
+
+    _this.handleMouseDown = function (e) {
+      _this.handleChange(e, true);
+      window.addEventListener('mousemove', _this.handleChange);
+      window.addEventListener('mouseup', _this.handleMouseUp);
+    };
+
+    _this.handleMouseUp = function () {
+      _this.unbindEventListeners();
+    };
+
+    _this.throttle = (0, _lodash2.default)(function (fn, data) {
       fn(data);
     }, 50);
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleMouseDown = this.handleMouseDown.bind(this);
-    this.handleMouseUp = this.handleMouseUp.bind(this);
+    return _this;
   }
 
   _createClass(Saturation, [{
@@ -66,60 +125,35 @@ var Saturation = (function (_ReactCSS$Component) {
       };
     }
   }, {
-    key: 'handleChange',
-    value: function handleChange(e) {
-      var container = React.findDOMNode(this.refs.container);
-      var containerWidth = container.clientWidth;
-      var containerHeight = container.clientHeight;
-      var left = e.pageX - (container.getBoundingClientRect().left + window.pageXOffset);
-      var top = e.pageY - (container.getBoundingClientRect().top + window.pageYOffset);
-
-      if (left < 0) {
-        left = 0;
-      } else if (left > containerWidth) {
-        left = containerWidth;
-      } else if (top < 0) {
-        top = 0;
-      } else if (top > containerHeight) {
-        top = containerHeight;
-      }
-
-      var saturation = left * 100 / containerWidth;
-      var bright = -(top * 100 / containerHeight) + 100;
-
-      this.throttle(this.props.onChange, { h: this.props.hsl.h, s: saturation, v: bright, a: this.props.hsl.a });
+    key: 'componentWillUnmount',
+    value: function componentWillUnmount() {
+      this.unbindEventListeners();
     }
   }, {
-    key: 'handleMouseDown',
-    value: function handleMouseDown(e) {
-      e.preventDefault();
-      this.handleChange(e);
-      window.addEventListener('mousemove', this.handleChange);
-      window.addEventListener('mouseup', this.handleMouseUp);
-    }
-  }, {
-    key: 'handleMouseUp',
-    value: function handleMouseUp() {
+    key: 'unbindEventListeners',
+    value: function unbindEventListeners() {
       window.removeEventListener('mousemove', this.handleChange);
       window.removeEventListener('mouseup', this.handleMouseUp);
     }
   }, {
     key: 'render',
     value: function render() {
-      var pointer = React.createElement('div', { style: this.styles().circle });
+      var pointer = _react2.default.createElement('div', { style: this.styles().circle });
 
       if (this.props.pointer) {
-        pointer = React.createElement(this.props.pointer, this.props);
+        pointer = _react2.default.createElement(this.props.pointer, this.props);
       }
 
-      return React.createElement(
+      return _react2.default.createElement(
         'div',
-        { style: this.styles().color, ref: 'container', onMouseDown: this.handleMouseDown },
-        React.createElement(
+        { style: this.styles().color, ref: 'container', onMouseDown: this.handleMouseDown,
+          onTouchMove: this.handleChange,
+          onTouchStart: this.handleChange },
+        _react2.default.createElement(
           'div',
           { style: this.styles().white },
-          React.createElement('div', { style: this.styles().black }),
-          React.createElement(
+          _react2.default.createElement('div', { style: this.styles().black }),
+          _react2.default.createElement(
             'div',
             { style: this.styles().pointer, ref: 'pointer' },
             pointer
@@ -130,6 +164,6 @@ var Saturation = (function (_ReactCSS$Component) {
   }]);
 
   return Saturation;
-})(ReactCSS.Component);
+}(_reactcss2.default.Component);
 
-module.exports = Saturation;
+exports.default = Saturation;
