@@ -1,45 +1,76 @@
 'use strict';
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.Photoshop = undefined;
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; desc = parent = getter = undefined; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+var _react = require('react');
 
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
+var _react2 = _interopRequireDefault(_react);
 
-function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+var _reactcss = require('reactcss');
 
-var React = require('react');
-var ReactCSS = require('reactcss');
+var _reactcss2 = _interopRequireDefault(_reactcss);
 
-var _require = require('../common');
+var _reactAddonsShallowCompare = require('react-addons-shallow-compare');
 
-var Saturation = _require.Saturation;
-var Hue = _require.Hue;
+var _reactAddonsShallowCompare2 = _interopRequireDefault(_reactAddonsShallowCompare);
 
-var PhotoshopFields = require('./PhotoshopFields');
-var PhotoshopPointerCircle = require('./PhotoshopPointerCircle');
-var PhotoshopPointer = require('./PhotoshopPointer');
+var _common = require('../common');
 
-var PhotoshopPicker = (function (_ReactCSS$Component) {
-  _inherits(PhotoshopPicker, _ReactCSS$Component);
+var _PhotoshopFields = require('./PhotoshopFields');
 
-  function PhotoshopPicker(props) {
-    _classCallCheck(this, PhotoshopPicker);
+var _PhotoshopFields2 = _interopRequireDefault(_PhotoshopFields);
 
-    _get(Object.getPrototypeOf(PhotoshopPicker.prototype), 'constructor', this).call(this);
+var _PhotoshopPointerCircle = require('./PhotoshopPointerCircle');
 
-    this.state = {
-      currentColor: props.hex
+var _PhotoshopPointerCircle2 = _interopRequireDefault(_PhotoshopPointerCircle);
+
+var _PhotoshopPointer = require('./PhotoshopPointer');
+
+var _PhotoshopPointer2 = _interopRequireDefault(_PhotoshopPointer);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var Photoshop = exports.Photoshop = function (_ReactCSS$Component) {
+  _inherits(Photoshop, _ReactCSS$Component);
+
+  function Photoshop(props) {
+    _classCallCheck(this, Photoshop);
+
+    var _this = _possibleConstructorReturn(this, Object.getPrototypeOf(Photoshop).call(this));
+
+    _this.shouldComponentUpdate = _reactAddonsShallowCompare2.default.bind(_this, _this, arguments[0], arguments[1]);
+
+    _this.handleChange = function (data) {
+      _this.props.onChange(data);
     };
 
-    this.handleChange = this.handleChange.bind(this);
-    this.handleAccept = this.handleAccept.bind(this);
-    this.handleCancel = this.handleCancel.bind(this);
+    _this.handleAccept = function () {
+      _this.props.onAccept && _this.props.onAccept();
+    };
+
+    _this.handleCancel = function () {
+      _this.props.onCancel && _this.props.onCancel();
+    };
+
+    _this.state = {
+      currentColor: props.hex
+    };
+    return _this;
   }
 
-  _createClass(PhotoshopPicker, [{
+  _createClass(Photoshop, [{
     key: 'classes',
     value: function classes() {
       return {
@@ -48,6 +79,7 @@ var PhotoshopPicker = (function (_ReactCSS$Component) {
             background: '#DCDCDC',
             borderRadius: '4px',
             boxShadow: '0 0 0 1px rgba(0,0,0,.25), 0 8px 16px rgba(0,0,0,.15)',
+            boxSizing: 'initial',
             width: '513px'
           },
           head: {
@@ -101,7 +133,7 @@ var PhotoshopPicker = (function (_ReactCSS$Component) {
             marginBottom: '2px',
             marginTop: '1px'
           },
-          'new': {
+          new: {
             height: '34px',
             background: 'rgb(' + this.props.rgb.r + ', ' + this.props.rgb.g + ', ' + this.props.rgb.b + ')',
             boxShadow: 'inset 1px 0 0 #000, inset -1px 0 0 #000, inset 0 1px 0 #000'
@@ -140,90 +172,75 @@ var PhotoshopPicker = (function (_ReactCSS$Component) {
       };
     }
   }, {
-    key: 'handleChange',
-    value: function handleChange(data) {
-      this.props.onChange(data);
-    }
-  }, {
-    key: 'handleAccept',
-    value: function handleAccept() {
-      this.props.onAccept && this.props.onAccept();
-    }
-  }, {
-    key: 'handleCancel',
-    value: function handleCancel() {
-      this.props.onCancel && this.props.onCancel();
-    }
-  }, {
     key: 'render',
     value: function render() {
       var header;
 
       if (this.props.header) {
-        header = React.createElement(
+        header = _react2.default.createElement(
           'div',
           { style: this.styles().head },
           this.props.header
         );
       }
 
-      return React.createElement(
+      return _react2.default.createElement(
         'div',
         { style: this.styles().picker },
         header,
-        React.createElement(
+        _react2.default.createElement(
           'div',
           { style: this.styles().body, className: 'flexbox-fix' },
-          React.createElement(
+          _react2.default.createElement(
             'div',
             { style: this.styles().saturation },
-            React.createElement(Saturation, _extends({}, this.styles().Saturation, this.props, { pointer: PhotoshopPointerCircle, onChange: this.handleChange }))
+            _react2.default.createElement(_common.Saturation, _extends({}, this.styles().Saturation, this.props, { pointer: _PhotoshopPointerCircle2.default, onChange: this.handleChange }))
           ),
-          React.createElement(
+          _react2.default.createElement(
             'div',
             { style: this.styles().hue },
-            React.createElement(Hue, _extends({}, this.styles().Hue, this.props, { pointer: PhotoshopPointer, onChange: this.handleChange }))
+            _react2.default.createElement(_common.Hue, _extends({}, this.styles().Hue, this.props, { pointer: _PhotoshopPointer2.default, onChange: this.handleChange }))
           ),
-          React.createElement(
+          _react2.default.createElement(
             'div',
             { style: this.styles().controls },
-            React.createElement(
+            _react2.default.createElement(
               'div',
               { style: this.styles().top, className: 'flexbox-fix' },
-              React.createElement(
+              _react2.default.createElement(
                 'div',
                 { style: this.styles().previews },
-                React.createElement(
+                _react2.default.createElement(
                   'div',
                   { style: this.styles().label },
                   'new'
                 ),
-                React.createElement(
+                _react2.default.createElement(
                   'div',
                   { style: this.styles().swatches },
-                  React.createElement('div', { style: this.styles()['new'] }),
-                  React.createElement('div', { style: this.styles().current })
+                  _react2.default.createElement('div', { style: this.styles().new }),
+                  _react2.default.createElement('div', { style: this.styles().current })
                 ),
-                React.createElement(
+                _react2.default.createElement(
                   'div',
                   { style: this.styles().label },
                   'current'
                 )
               ),
-              React.createElement(
+              _react2.default.createElement(
                 'div',
                 { style: this.styles().actions },
-                React.createElement(
+                _react2.default.createElement(
                   'div',
-                  { style: this.styles().acceptButton, onClick: this.handleAccept },
+                  { style: this.styles().acceptButton, ref: 'accept', onClick: this.handleAccept },
                   'OK'
                 ),
-                React.createElement(
+                _react2.default.createElement(
                   'div',
-                  { style: this.styles().button, onClick: this.handleCancel },
+                  { style: this.styles().button, ref: 'cancel', onClick: this.handleCancel },
                   'Cancel'
                 ),
-                React.createElement(PhotoshopFields, this.props)
+                _react2.default.createElement(_PhotoshopFields2.default, this.props)
               )
             )
           )
@@ -232,11 +249,11 @@ var PhotoshopPicker = (function (_ReactCSS$Component) {
     }
   }]);
 
-  return PhotoshopPicker;
-})(ReactCSS.Component);
+  return Photoshop;
+}(_reactcss2.default.Component);
 
-PhotoshopPicker.defaultProps = {
+Photoshop.defaultProps = {
   header: 'Color Picker'
 };
 
-module.exports = PhotoshopPicker;
+exports.default = (0, _common.ColorWrap)(Photoshop);
