@@ -1,5 +1,18 @@
+'use strict';
 
-module.exports = {
-  Container: require('./components/Container.jsx'),
-  Grid: require('./components/Grid.jsx'),
-};
+Object.defineProperty(exports, '__esModule', {
+  value: true,
+});
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+
+var _libComponentsContainer = require('./lib/components/Container');
+
+var _libComponentsContainer2 = _interopRequireDefault(_libComponentsContainer);
+
+var _libComponentsGrid = require('./lib/components/Grid');
+
+var _libComponentsGrid2 = _interopRequireDefault(_libComponentsGrid);
+
+exports.Container = _libComponentsContainer2['default'];
+exports.Grid = _libComponentsGrid2['default'];
