@@ -1,2 +1,14 @@
+'use strict';
 
-module.exports = require('./components/Docs.jsx');
+Object.defineProperty(exports, '__esModule', {
+  value: true,
+});
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+
+var _libComponentsDocs = require('./lib/components/Docs');
+
+var _libComponentsDocs2 = _interopRequireDefault(_libComponentsDocs);
+
+exports['default'] = _libComponentsDocs2['default'];
+module.exports = exports['default'];
