@@ -1,6 +1,23 @@
-module.exports = {
-  Raised: require('./src/components/Raised.jsx'),
-  Tile: require('./src/components/Tile.jsx'),
+'use strict';
 
-  Tabs: require('./src/components/Tabs.jsx'),
-};
+Object.defineProperty(exports, '__esModule', {
+  value: true,
+});
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+
+var _libComponentsRaised = require('./lib/components/Raised');
+
+var _libComponentsRaised2 = _interopRequireDefault(_libComponentsRaised);
+
+var _libComponentsTile = require('./lib/components/Tile');
+
+var _libComponentsTile2 = _interopRequireDefault(_libComponentsTile);
+
+var _libComponentsTabs = require('./lib/components/Tabs');
+
+var _libComponentsTabs2 = _interopRequireDefault(_libComponentsTabs);
+
+exports.Raised = _libComponentsRaised2['default'];
+exports.Tile = _libComponentsTile2['default'];
+exports.Tabs = _libComponentsTabs2['default'];
