@@ -1,8 +1,6 @@
-
-module.exports = {
-  Alpha: require('./Alpha'),
-  Checkboard: require('./Checkboard'),
-  EditableInput: require('./EditableInput'),
-  Hue: require('./Hue'),
-  Saturation: require('./Saturation'),
-};
+export { default as Alpha } from './Alpha'
+export { default as Checkboard } from './Checkboard'
+export { default as EditableInput } from './EditableInput'
+export { default as Hue } from './Hue'
+export { default as Saturation } from './Saturation'
+export { default as ColorWrap } from './ColorWrap'
